Moving project to idam package
