+++ src/main/java/uk/gov/hmcts/reform/idam/config/AppInsightsConfiguration.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.reform.idam.config;
+
+import com.microsoft.applicationinsights.TelemetryClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppInsightsConfiguration {
+    @Bean
+    public TelemetryClient telemetryClient() {
+        return new TelemetryClient();
+    }
+}
