+++ src/smokeTest/java/uk/gov/hmcts/reform/idam/controllers/SampleSmokeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.controllers;
+package uk.gov.hmcts.reform.idam.controllers;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -35,5 +35,5 @@ class SampleSmokeTest {
 
         Assertions.assertEquals(200, response.statusCode());
         Assertions.assertTrue(response.asString().startsWith("Welcome"));
-    } 
+    }
 }
